refactor: AI client 패키지 이동
- 기존의 global/client에서 domain/chat/client로 이동
