+++ chatbot/backend/src/main/java/com/hallachatbot/backend/domain/chat/client/dto/AiChatRequest.java
@@ -1,4 +1,4 @@
-package com.hallachatbot.backend.global.client.dto;
+package com.hallachatbot.backend.domain.chat.client.dto;
 
 import java.util.List;
