+++ chatbot/backend/src/main/java/com/hallachatbot/backend/domain/chat/component/ChatStreamHandler.java
@@ -7,7 +7,7 @@ import org.springframework.http.codec.ServerSentEvent;
 import org.springframework.stereotype.Component;
 
 import com.hallachatbot.backend.domain.chat.service.ChatStreamContext;
-import com.hallachatbot.backend.global.client.dto.AiServiceResponse;
+import com.hallachatbot.backend.domain.chat.client.dto.AiServiceResponse;
 import com.hallachatbot.backend.global.sse.SseEventFactory;
 
 import lombok.RequiredArgsConstructor;
