+++ chatbot/backend/src/test/java/com/hallachatbot/backend/domain/chat/component/ChatStreamHandlerTest.java
@@ -19,7 +19,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.codec.ServerSentEvent;
 
 import com.hallachatbot.backend.domain.chat.service.ChatStreamContext;
-import com.hallachatbot.backend.global.client.dto.AiServiceResponse;
+import com.hallachatbot.backend.domain.chat.client.dto.AiServiceResponse;
 import com.hallachatbot.backend.global.sse.SseEventFactory;
 
 @ExtendWith(MockitoExtension.class)
