+++ chatbot/backend/src/test/java/com/hallachatbot/backend/domain/chat/service/ChatServiceTest.java
@@ -31,8 +31,8 @@ import com.hallachatbot.backend.domain.chat.component.ChatStreamHandler;
 import com.hallachatbot.backend.domain.chat.component.ChatWriter;
 import com.hallachatbot.backend.domain.chat.dto.request.ChatRequest;
 import com.hallachatbot.backend.domain.usage.service.UsageService;
-import com.hallachatbot.backend.global.client.dto.AiServiceResponse;
-import com.hallachatbot.backend.global.client.service.AiServiceClient;
+import com.hallachatbot.backend.domain.chat.client.dto.AiServiceResponse;
+import com.hallachatbot.backend.domain.chat.client.service.AiServiceClient;
 import com.hallachatbot.backend.global.errorcode.UsageErrorCode;
 import com.hallachatbot.backend.global.exception.ChatStreamErrorHandler;
 import com.hallachatbot.backend.global.exception.UsageException;
