+++ chatbot/backend/src/test/java/com/hallachatbot/backend/global/client/service/AiServiceClientTest.java
@@ -16,8 +16,10 @@ import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.hallachatbot.backend.domain.chat.client.service.AiServiceClient;
+import com.hallachatbot.backend.domain.chat.client.service.AiServiceClientImpl;
 import com.hallachatbot.backend.domain.chat.dto.request.ChatRequest;
-import com.hallachatbot.backend.global.client.dto.AiServiceResponse;
+import com.hallachatbot.backend.domain.chat.client.dto.AiServiceResponse;
 
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
